task_3/requirements.txt
# Telegram parsing
telethon==1.36.0
python-dotenv==1.0.1

# Data
pandas==2.2.3
pyarrow==17.0.0
numpy==1.26.4

# Embeddings
sentence-transformers==3.3.1
transformers==4.46.3
torch==2.4.1
nltk==3.9.1
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142

# Clustering
scikit-learn==1.4.2
hdbscan==0.8.39
umap-learn==0.5.6
bertopic==0.16.4

# Sentiment
sentencepiece==0.2.0

# Visualization
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Utilities
click==8.1.7
tqdm==4.66.5